fix warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4626 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/wpc/core.c
@@ -23,7 +23,7 @@
 
  #include "gts3dmd.h"
  UINT8  g_raw_gtswpc_dmd[GTS3DMD_FRAMES_5C*0x200];
- UINT8  g_raw_gtswpc_dmdframes = 0;
+ UINT32 g_raw_gtswpc_dmdframes = 0;
 
  UINT32 g_needs_DMD_update = 1;
 #endif
